+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.8.1] - 2018-XX-XX
+### Added
+- [Search] History
+
 ### Fixed
 - [Bug] Save vote-data in local keychain [#322](https://github.com/demokratie-live/democracy-client/issues/322)